electron_project/python_mcp/requirements.txt
dedalus-labs
python-dotenv
asyncio
